requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
APScheduler==3.11.0
PyYAML==6.0.2
